Decoupling Capacitors
This step involved making the capacitors that filter the electrical noise from the microcrontroller. The official documentation says to use 0.1uF by each VCC, which would make 4, and a 10uF overall. But when looking at the Teensy Schematic, it uses 3 0.1uF and 1 1uF. The guide says to combine both, and it still works functionally. so that is what has been added here.