Footprinting the Schematic
This is done to create the associations of the components with the actual schematic that we have made. The software can do this on it's own with smaller projects, but in future do all of this manually. Whether that be during the project or at the end of the schematic build.